1) Moved the FluidSynth code to its own specific source file. 2) A bit of re-structuring performed in-order to move the stop/play calls into its own dedicated thread. 3) Further tweaks have been made to avoid a rare crash during music override situations (the nature of this specific crash only had an effect on the FluidSynth thread and consequently allowed the game to keep running).
