+++ sc2kfix.vcxproj.filters
@@ -488,6 +488,9 @@
     <ClCompile Include="hooks\hook_scenario.cpp">
       <Filter>Source Files\Hooks</Filter>
     </ClCompile>
+    <ClCompile Include="modules\fluidsynth_engine.cpp">
+      <Filter>Source Files\Modules</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="sc2kfix.rc">
